+++ client/app/article/[id]/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+export default function Article(props: any) {
+    return (
+        <p>article id: {props.params.id}</p>
+    )
+}
