--- client/app/login/layout.tsx
@@ -1,28 +1,0 @@
-import Footer from "@/components/footer";
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-
-
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "AI-NES",
-  description: "OSS Project",
-};
-
-export default function LoginLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        {children}
-      </body>
-    </html>
-  );
-}
-
-
