+++ client/components/navbar.tsx
@@ -1,8 +1,9 @@
 'use client'
 
-import { Fragment } from 'react'
+import { Fragment, useState } from 'react'
 import { Disclosure, Menu, Transition } from '@headlessui/react'
 import { Bars3Icon, BellIcon, XMarkIcon, UserCircleIcon } from '@heroicons/react/24/outline'
+import LoginModal from "./loginmodal";
 
 const navigation = [
   { name: '대시보드', href: './', current: true },
@@ -16,6 +17,9 @@ function classNames(...classes: string[]) {
 }
 
 export default function Navbar() {
+
+  const [isLoginOpen, setIsLoginOpen] = useState(false);
+
   return (
     <Disclosure as="nav" className="bg-gray-800">
       {({ open }) => (
@@ -76,7 +80,7 @@ export default function Navbar() {
                     <Menu.Button className="relative flex rounded-full bg-gray-800 text-sm focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800">
                       <span className="absolute -inset-1.5" />
                       <span className="sr-only">Open user menu</span>
-                      <UserCircleIcon className='h-8 w-8 text-gray-400 hover:text-white'/>
+                      <UserCircleIcon className='h-8 w-8 text-gray-400 hover:text-white' />
                     </Menu.Button>
                   </div>
                   <Transition
@@ -91,12 +95,12 @@ export default function Navbar() {
                     <Menu.Items className="absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
                       <Menu.Item>
                         {({ active }) => (
-                          <a
-                            href="#"
-                            className={classNames(active ? 'bg-gray-100' : '', 'block px-4 py-2 text-sm text-gray-700')}
+                          <button
+                            onClick={() => setIsLoginOpen(true)}
+                            className={classNames(active ? 'bg-gray-100' : '', 'text-left block px-4 py-2 text-sm text-gray-700 w-full')}
                           >
                             Sign in
-                          </a>
+                          </button>
                         )}
                       </Menu.Item>
                     </Menu.Items>
@@ -124,8 +128,10 @@ export default function Navbar() {
               ))}
             </div>
           </Disclosure.Panel>
+          <LoginModal open={isLoginOpen} setOpen={setIsLoginOpen}/>
         </>
       )}
+
     </Disclosure>
   )
 }
